Exclude probowl week from imports
The probowl week does not need to be included in imports because the
probowl is not part of our fantasy league.  In order to exclude the
probowl week, we only include weeks that have games with a team that
exists in the database.  The teams 'AFC' and 'NFC' do not exist in the
Lightdatasys database, so the week for the Probowl game is excluded.